+++ scheduling/src/main/java/com/nabla/sdk/scheduling/scene/appointments/AppointmentsContentFragment.kt
@@ -65,7 +65,7 @@ internal class AppointmentsContentFragment : SchedulingBaseFragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View {
-        return NablaSchedulingFragmentAppointmentsContentBinding.inflate(inflater, container, false).root
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {